restructure some classes
