+++ yabte/backtest/order.py
@@ -1,10 +1,11 @@
 from __future__ import annotations
 
 import logging
+from collections import Counter, deque
 from dataclasses import dataclass, field
 from decimal import Decimal
 from enum import Enum
-from typing import Dict, List, Optional, Tuple
+from typing import Dict, Iterable, List, Optional, Tuple
 
 import pandas as pd
 from mypy_extensions import mypyc_attr
@@ -102,6 +103,56 @@ class Order:
         """Applies order to `self.book` for time `ts` using provided `day_data` and
         dictionary of asset information `asset_map`."""
         raise NotImplementedError("The apply methods needs to be implemented.")
+
+
+class Orders:
+    """Double ended queue of orders."""
+
+    def __init__(self):
+        self.deque = deque()
+
+    def __len__(self):
+        return len(self.deque)
+
+    def __iter__(self):
+        return iter(self.deque)
+
+    def popleft(self):
+        return self.deque.popleft()
+
+    def append(self, order: Order):
+        return self.deque.append(order)
+
+    def extend(self, orders: Iterable[Order]):
+        return self.deque.extend(orders)
+
+    def sort_by_priority(self):
+        """Sorts orders by order priority."""
+        ou_sorted = sorted(self.deque, key=lambda o: o.priority, reverse=True)
+        self.deque.clear()
+        self.deque.extend(ou_sorted)
+
+    def remove_duplicate_keys(self) -> List[Order]:
+        """Remove older orders with same key.
+
+        Returns a list of orders than were removed with status set to REPLACED.
+        """
+        removed = []
+        cntr = Counter(o.key for o in self.deque if o.key is not None)
+        if any(v > 1 for v in cntr.values()):
+            kept = []
+            while self.deque:
+                o = self.deque.popleft()
+                if o.key in cntr and cntr[o.key] > 1:
+                    o.status = OrderStatus.REPLACED
+                    removed.append(o)
+                    cntr[o.key] -= 1
+                else:
+                    kept.append(o)
+            self.deque.clear()
+            self.deque.extend(kept)
+
+        return removed
 
 
 @mypyc_attr(allow_interpreted_subclasses=True)
